Use official vault gem to handle IAM auth
The official 'vault' gem from Hashicorp has a method to handle IAM
authentication. The arguments are a little wonky, requiring some
overly-explicit arguments to provide the IAM credentials and the
non-default URI path to the AWS auth provider, but it works.

Using this method also handles using additional headers in the
pre-signed request.

Also, bumps VERSION to 0.0.1.